requirements.txt
certifi==2020.6.20
cffi==1.11.5
cycler==0.10.0
decorator==4.4.2
imageio==2.8.0
kiwisolver==1.2.0
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
Pillow==7.1.2
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.0
six==1.15.0
tifffile==2020.6.3
